% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scooldr-package.R
\docType{package}
\name{scooldr-package}
\alias{scooldr}
\alias{scooldr-package}
\title{scooldr: Interact with Scoold API}
\description{
Allow admin to interact with his Scoold app from R.
}
\author{
\strong{Maintainer}: Cervan Girard \email{cervan.girard@gmail.com}

}
\keyword{internal}
